Implement saving of opened tabs
Fixes https://github.com/rugk/how-did-i-get-here/issues/1

It does assign an unique number to each tab and records the parent tab data (url, favicon, etc.) for it in two places: the parent data and the background script.

When building the UI, it always prefers the tab data and falls back to querying the background script.

When displaying the tab data it prefers the original tab title, etc, but as for status information (tab hidden, etc.) of course the up-to-date info is used.